# speed: .is_empty_object