Follow up on comments from here: https://github.com/easystats/insight/pull/484#

The new flag/indexing strategy improves the speed of `get_data` by 100x in a naive benchmark I just ran:

New:

```
  expression         min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
  <bch:expr>    <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>
1 get_data(mod)   59.3ms   78.6ms      13.8    57.3MB        0     3     0
```


Old:

```
  expression      min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time
  <bch:expr>    <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm>
1 get_data(mod) 9.61s  9.61s     0.104     235MB     1.14     1    11      9.61s
```